BUGFIX: Correctly handle shell argument
The `if` condition did worked only by coincidence - it did run whatever
was in `$1` in a subshell (see Compound Commands in bash(1)) and most
likely printed out an error like

  ./proj_gen.sh: line 5: 69: command not found

so the condition was not true and the `DEF` has remained set to `$1`.
Unless somebody happened to pass some actual shell command in the
argument..

If the `$1` was empty, the condition did succeed by running a subshell
with no command, so the `DEF` was set to the default value of `1`.